fix(core): clear remaining AppleClang warnings (tests + template-only) and mark Boost/Eigen SYSTEM
The library-only build missed warnings in templates that are only
instantiated by the test/binding TUs. Building the full tree surfaces them;
fix the rest of the first-party core warnings at the root cause:

- -Wunused-lambda-capture: drop unused [this] from the remaining Boost.Spirit
  parser semantic-action lambdas (settings_parsers/{algorithm,endgames,
  tracking,random}) that only the test build instantiates
- -Wsign-conversion: explicit static_cast at boundaries in observers.hpp
  (std::setprecision takes int) and across the affected test files
- -Wsign-compare / -Wimplicit-int-float-conversion / -Wfloat-conversion:
  explicit casts in the test files
- -Wunused-parameter / -Wunused-variable / -Wunused-but-set-variable: drop
  unused names/locals (trace.hpp, several tests)

Also mark Boost and Eigen include directories as SYSTEM so any warning that
originates inside those third-party headers stays out of our build output
(GMP/MPFR/MPC were already SYSTEM). Fixes a keyword-order bug in the bindings
CMake ("PUBLIC SYSTEM" -> "SYSTEM PUBLIC") that left Boost/Python/NumPy not
actually treated as system.

Core (library + tests) now compiles warning-free; all 10 C++ suites pass.
Only the python bindings still emit warnings (handled next, with -Wno-conversion
removed).

+++ core/test/classes/function_tree_transform.cpp
@@ -289,7 +289,7 @@ BOOST_AUTO_TEST_CASE(complicated)
 	auto x = Variable::Make("x");
 	auto y = Variable::Make("y");
 
-	auto n = (((((2*x*1)*y)+(0*(pow(x,2))))/2)-(0*((pow(x,2))*y)/(pow(2,2))));
+	auto n = (((((2*x*1)*y)+(0*(pow(x,2))))/2)-(0*((pow(x,2))*y)/(static_cast<int>(pow(2,2)))));
 
 	auto ns = bertini::Simplify(n);
